correção da distribuição de tropas feita pela IA
Basicamente aleatório.
